# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Il magazzino del man
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type AuthUserRole = typeof AuthUserRole[keyof typeof AuthUserRole];


export const AuthUserRole = {
  admin: 'admin',
  client: 'client',
} as const;

export interface AuthUser {
  id: number;
  email: string;
  role: AuthUserRole;
  /** @nullable */
  name?: string | null;
}

export type ProductCondition = typeof ProductCondition[keyof typeof ProductCondition];


export const ProductCondition = {
  new_with_tags: 'new_with_tags',
  like_new: 'like_new',
  used: 'used',
} as const;

/**
 * @nullable
 */
export type ProductCommissionType = typeof ProductCommissionType[keyof typeof ProductCommissionType] | null;


export const ProductCommissionType = {
  percentage: 'percentage',
  fixed: 'fixed',
} as const;

export interface Product {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  condition: ProductCondition;
  /** @nullable */
  brand?: string | null;
  /** @nullable */
  color?: string | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  inStock: boolean;
  /** @nullable */
  commissionType?: ProductCommissionType;
  /** @nullable */
  commissionValue?: number | null;
  /** @nullable */
  netProfit?: number | null;
  createdAt: string;
}

export type ProductInputCondition = typeof ProductInputCondition[keyof typeof ProductInputCondition];


export const ProductInputCondition = {
  new_with_tags: 'new_with_tags',
  like_new: 'like_new',
  used: 'used',
} as const;

export type ProductInputCommissionType = typeof ProductInputCommissionType[keyof typeof ProductInputCommissionType];


export const ProductInputCommissionType = {
  percentage: 'percentage',
  fixed: 'fixed',
} as const;

export interface ProductInput {
  title: string;
  description?: string;
  price: number;
  originalPrice?: number;
  condition: ProductInputCondition;
  brand?: string;
  color?: string;
  category?: string;
  imageUrl?: string;
  commissionType?: ProductInputCommissionType;
  commissionValue?: number;
}

export type ProductUpdateCondition = typeof ProductUpdateCondition[keyof typeof ProductUpdateCondition];


export const ProductUpdateCondition = {
  new_with_tags: 'new_with_tags',
  like_new: 'like_new',
  used: 'used',
} as const;

export type ProductUpdateCommissionType = typeof ProductUpdateCommissionType[keyof typeof ProductUpdateCommissionType];


export const ProductUpdateCommissionType = {
  percentage: 'percentage',
  fixed: 'fixed',
} as const;

export interface ProductUpdate {
  title?: string;
  description?: string;
  price?: number;
  originalPrice?: number;
  condition?: ProductUpdateCondition;
  brand?: string;
  color?: string;
  category?: string;
  imageUrl?: string;
  inStock?: boolean;
  commissionType?: ProductUpdateCommissionType;
  commissionValue?: number;
}

export type ProductSummaryCategoriesItem = {
  category: string;
  count: number;
};

export interface ProductSummary {
  total: number;
  inStock: number;
  sold: number;
  categories: ProductSummaryCategoriesItem[];
  totalValue: number;
}

export type ProposalStatus = typeof ProposalStatus[keyof typeof ProposalStatus];


export const ProposalStatus = {
  pending: 'pending',
  accepted: 'accepted',
  rejected: 'rejected',
} as const;

/**
 * @nullable
 */
export type ProposalCommissionType = typeof ProposalCommissionType[keyof typeof ProposalCommissionType] | null;


export const ProposalCommissionType = {
  percentage: 'percentage',
  fixed: 'fixed',
} as const;

export interface Proposal {
  id: number;
  itemName: string;
  category: string;
  desiredPayout: number;
  /** @nullable */
  imageUrl?: string | null;
  status: ProposalStatus;
  submitterEmail: string;
  /** @nullable */
  submitterName?: string | null;
  /** @nullable */
  commissionType?: ProposalCommissionType;
  /** @nullable */
  commissionValue?: number | null;
  /** @nullable */
  netProfit?: number | null;
  /** @nullable */
  adminNotes?: string | null;
  createdAt: string;
}

export interface ProposalInput {
  itemName: string;
  category: string;
  desiredPayout: number;
  imageUrl?: string;
  submitterEmail: string;
  submitterName?: string;
}

export type CommissionInputCommissionType = typeof CommissionInputCommissionType[keyof typeof CommissionInputCommissionType];


export const CommissionInputCommissionType = {
  percentage: 'percentage',
  fixed: 'fixed',
} as const;

export interface CommissionInput {
  commissionType: CommissionInputCommissionType;
  commissionValue: number;
  adminNotes?: string;
}

export type ChatMessageSenderRole = typeof ChatMessageSenderRole[keyof typeof ChatMessageSenderRole];


export const ChatMessageSenderRole = {
  admin: 'admin',
  client: 'client',
} as const;

export interface ChatMessage {
  id: number;
  proposalId: number;
  senderRole: ChatMessageSenderRole;
  /** @nullable */
  senderName?: string | null;
  content: string;
  createdAt: string;
}

export type ChatMessageInputSenderRole = typeof ChatMessageInputSenderRole[keyof typeof ChatMessageInputSenderRole];


export const ChatMessageInputSenderRole = {
  admin: 'admin',
  client: 'client',
} as const;

export interface ChatMessageInput {
  content: string;
  senderRole: ChatMessageInputSenderRole;
  senderName?: string;
}

export interface CheckoutInput {
  productId: number;
  cardNumber: string;
  expiryMonth: number;
  expiryYear: number;
  cvc: string;
  /** Amount in cents (e.g. 1000 = €10.00) */
  amountCents: number;
  /** Buyer email address — invoice will be sent here */
  buyerEmail?: string;
  /** Buyer display name for the invoice */
  buyerName?: string;
}

export interface CheckoutResult {
  success: boolean;
  transactionId: string;
  amountCents: number;
  message?: string;
}

export type FinancialSummaryMonthlyDataItem = {
  month: string;
  revenue: number;
  expenses: number;
  profit: number;
};

export interface FinancialSummary {
  totalRevenue: number;
  totalExpenses: number;
  totalProfit: number;
  monthlyData: FinancialSummaryMonthlyDataItem[];
}

export interface FinancialItem {
  id: number;
  productTitle: string;
  salePrice: number;
  commissionType: string;
  commissionValue: number;
  netProfit: number;
  soldAt: string;
}

export type CmsContentSocialLinksItem = {
  label: string;
  url: string;
  icon: string;
};

export interface CmsContent {
  heroHeadline: string;
  heroSubheadline: string;
  chiSiamoTitle: string;
  chiSiamoBody: string;
  founderName: string;
  socialLinks: CmsContentSocialLinksItem[];
}

export type CmsContentUpdateSocialLinksItem = {
  label: string;
  url: string;
  icon: string;
};

export interface CmsContentUpdate {
  heroHeadline?: string;
  heroSubheadline?: string;
  chiSiamoTitle?: string;
  chiSiamoBody?: string;
  founderName?: string;
  socialLinks?: CmsContentUpdateSocialLinksItem[];
}

export type ListProductsParams = {
category?: string;
condition?: ListProductsCondition;
brand?: string;
color?: string;
minPrice?: number;
maxPrice?: number;
inStock?: boolean;
};

export type ListProductsCondition = typeof ListProductsCondition[keyof typeof ListProductsCondition];


export const ListProductsCondition = {
  new_with_tags: 'new_with_tags',
  like_new: 'like_new',
  used: 'used',
} as const;

export type ListProposalsParams = {
status?: ListProposalsStatus;
};

export type ListProposalsStatus = typeof ListProposalsStatus[keyof typeof ListProposalsStatus];


export const ListProposalsStatus = {
  pending: 'pending',
  accepted: 'accepted',
  rejected: 'rejected',
} as const;