# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for Il magazzino del man
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional filters
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: condition
          schema:
            type: string
            enum: [new_with_tags, like_new, used]
        - in: query
          name: brand
          schema:
            type: string
        - in: query
          name: color
          schema:
            type: string
        - in: query
          name: minPrice
          schema:
            type: number
        - in: query
          name: maxPrice
          schema:
            type: number
        - in: query
          name: inStock
          schema:
            type: boolean
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a new product (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /proposals/{id}/messages:
    get:
      operationId: listMessages
      tags: [chat]
      summary: Get chat messages for a proposal
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a chat message for a proposal
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "201":
          description: Sent message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
  /checkout/simulate:
    post:
      operationId: simulateCheckout
      tags: [checkout]
      summary: Simulate Stripe checkout (marks item as sold)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "200":
          description: Checkout result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutResult"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required: [success]
...
    Product:
      type: object
      required: [id, title, price, condition, inStock, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        originalPrice:
          type: ["number", "null"]
        condition:
          type: string
          enum: [new_with_tags, like_new, used]
        brand:
          type: ["string", "null"]
        color:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        inStock:
          type: boolean
        commissionType:
          type: ["string", "null"]
          enum: [percentage, fixed, null]
        commissionValue:
          type: ["number", "null"]
        netProfit:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
    ProductInput:
      type: object
      required: [title, price, condition]
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        condition:
          type: string
          enum: [new_with_tags, like_new, used]
        brand:
          type: string
        color:
          type: string
        category:
          type: string
        imageUrl:
          type: string
        commissionType:
          type: string
          enum: [percentage, fixed]
        commissionValue:
          type: number
    ProductUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        condition:
          type: string
          enum: [new_with_tags, like_new, used]
        brand:
          type: string
        color:
          type: string
        category:
          type: string
        imageUrl:
          type: string
        inStock:
          type: boolean
        commissionType:
          type: string
          enum: [percentage, fixed]
        commissionValue:
          type: number
...
    ChatMessage:
      type: object
      required: [id, proposalId, senderRole, content, createdAt]
      properties:
        id:
          type: integer
        proposalId:
          type: integer
        senderRole:
          type: string
          enum: [admin, client]
        senderName:
          type: ["string", "null"]
        content:
          type: string
        createdAt:
          type: string
          format: date-time
    ChatMessageInput:
      type: object
      required: [content, senderRole]
      properties:
        content:
          type: string
        senderRole:
          type: string
          enum: [admin, client]
        senderName:
          type: string
    CheckoutInput:
      type: object
      required: [productId, cardNumber, expiryMonth, expiryYear, cvc, amountCents]
      properties:
        productId:
          type: integer
        cardNumber:
          type: string
        expiryMonth:
          type: integer
        expiryYear:
          type: integer
        cvc:
          type: string
        amountCents:
          type: integer
          description: Amount in cents (e.g. 1000 = €10.00)
        buyerEmail:
          type: string
          format: email
          description: Buyer email address — invoice will be sent here
        buyerName:
          type: string
          description: Buyer display name for the invoice
    CheckoutResult:
      type: object
      required: [success, transactionId, amountCents]
      properties:
        success:
          type: boolean
        transactionId:
          type: string
        amountCents:
          type: integer
        message:
          type: string
...