# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Il magazzino del man
 * OpenAPI spec version: 0.1.0
 */
import type { ProductInputCommissionType } from './productInputCommissionType';
import type { ProductInputCondition } from './productInputCondition';

export interface ProductInput {
  title: string;
  description?: string;
  price: number;
  originalPrice?: number;
  condition: ProductInputCondition;
  brand?: string;
  color?: string;
  category?: string;
  imageUrl?: string;
  commissionType?: ProductInputCommissionType;
  commissionValue?: number;
}