# lib/api-zod/src/generated/types/successResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification for Il magazzino del man
 * OpenAPI spec version: 0.1.0
 */

export interface SuccessResponse {
  success: boolean;
}